move stats structs to a sub-package
