+++ ecs/stats/stats_test.go
@@ -1,0 +1,24 @@
+package stats
+
+import (
+	"fmt"
+	"reflect"
+	"testing"
+)
+
+func TestStats(t *testing.T) {
+	stats := WorldStats{
+		Entities:       EntityStats{},
+		ComponentCount: 1,
+		ComponentTypes: []reflect.Type{reflect.TypeOf(1)},
+		Locked:         false,
+		Archetypes: []ArchetypeStats{{
+			Size:           1,
+			Capacity:       128,
+			Components:     1,
+			ComponentIDs:   []uint8{0},
+			ComponentTypes: []reflect.Type{reflect.TypeOf(1)},
+		}},
+	}
+	fmt.Println(stats.String())
+}
