--- ecs/stats_test.go
@@ -1,24 +1,0 @@
-package ecs
-
-import (
-	"fmt"
-	"testing"
-)
-
-func TestWorldStats(t *testing.T) {
-	w := NewWorld()
-
-	posID := ComponentID[position](&w)
-	rotID := ComponentID[rotation](&w)
-
-	e0 := w.NewEntity()
-	e1 := w.NewEntity()
-	e2 := w.NewEntity()
-
-	w.Add(e0, posID)
-	w.Add(e1, posID, rotID)
-	w.Add(e2, posID, rotID)
-
-	stats := w.Stats()
-	fmt.Println(stats.String())
-}
